refactor oncoprint tooltip code
